--- a/activesupport/lib/active_support/core_ext/object/json.rb
+++ b/activesupport/lib/active_support/core_ext/object/json.rb
@@ -165,7 +165,8 @@ end
 class Array
   def as_json(options = nil) # :nodoc:
     if options
-      map { |v| v.as_json(options.dup) }
+      options = options.dup.freeze unless options.frozen?
+      map { |v| v.as_json(options) }
     else
       map { |v| v.as_json }
     end
@@ -189,7 +190,8 @@ class Hash
 
     result = {}
     if options
-      subset.each { |k, v| result[k.to_s] = v.as_json(options.dup) }
+      options = options.dup.freeze unless options.frozen?
+      subset.each { |k, v| result[k.to_s] = v.as_json(options) }
     else
       subset.each { |k, v| result[k.to_s] = v.as_json }
     end