Previously for every element in an Array and every value in a Hash we would duplicate the options hash passed in. This PR avoids the duplication.

I did this very conservatively, duping and freezing the Hash when it is not already frozen so that any cases mutating the Hash should be caught. I don't feel very strongly about this and we could remove this in the future.

This should improve `to_json` performance when options are passed in (passing options is generally slower but sometimes unavoidable).

https://gist.github.com/jhawthorn/ca9e2eb38a2a982d7e30f9ab315f89a5


**Before**

```
❯ ruby benchmark_as_json.rb
ruby 3.3.2 (2024-05-30 revision e5a195edf6) [arm64-darwin23]
Calculating -------------------------------------
obj.to_json(foo: :bar)
                         35.401k (± 0.7%) i/s -    177.700k in   5.019858s
obj.as_json(foo: :bar)
                         60.774k (± 0.5%) i/s -    304.250k in   5.006362s
```

**After**

```
❯ ruby benchmark_as_json.rb
ruby 3.3.2 (2024-05-30 revision e5a195edf6) [arm64-darwin23]
Calculating -------------------------------------
obj.to_json(foo: :bar)
                         61.766k (± 0.7%) i/s -    310.250k in   5.023192s
obj.as_json(foo: :bar)
                        226.693k (± 0.8%) i/s -      1.138M in   5.018156s
```